refactor(viewer) splitting files and removing unused ones
